Sber/requirements.txt
aiohttp==3.9.5
python-docx==1.1.0
openpyxl==3.1.2
requests==2.32.3
